Created object sprite sheet and more
I created the sprite sheet for objects, as well as tried to add some
clouds but those clouds are basically a jumbly mess of white cotton
balls appearing on the screen. Im going to try to work on shooting for
the enemy but it the easiest way is ray casting and I don't know how to
do that, so we will see.