Tried fixing the bullets
The bullets are not rendering even though the sprite is there and
rendering, I tried to do what I did for the player and enemy which was
remove the bullet and then add it back. However, that did not work, I
will have to look further into this bug. For now, it is a problem for a
different time.