Unknown bug
I have encountered a bug, I do not know why. Let me explain it, so you
start up the game, you go through the menu, it is perfectly fine.
However, when you get to actually playing the game, after about 30 or so
seconds, everything stops drawing. The render method inside of the
renderer is getting called, so I will have to debug this 700+ lined
jungle of code.