Worked on multiplayer
I worked a bit on the multiplayer feature. However, I can't seem to get
it to work without crashing. I think I might have to rewrite this entire
feature because it does not seem to be working, it might be that I'm
overloading the network because maybe the data is too big, so I will try
to find an easier and faster way to do this.