Improved multiplayer
Im working on so that people can have more people in their server.
However, that does require some logic.
I've also improved multiplayer so now instead of loading the enemy's
sprite over and over again, it's stored in a global variable so I can
just check if it's null, if it is then load it, if it isn't then don't
load it which improves the speed of multiplayer rendering.