There is a bug
There is a health bug in multiplayer, I tried to fix it but I can't
because it's impossible for me to clean my glasses because some how
instead of the stuff I'm trying to clean off of it it spreads. I'm going
to try to fix this bug when I'm able to see through my glasses.